Finalized code refactoring
