+++ .gitignore
@@ -1,2 +1,3 @@
 /Github-ignore/
 /venv/
+/optimized_datasets/processed_h5/
