+++ .idea/CS467-Top-n-Music-Genre-Classification-Neural-Network.iml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="PYTHON_MODULE" version="4">
+  <component name="NewModuleRootManager">
+    <content url="file://$MODULE_DIR$">
+      <excludeFolder url="file://$MODULE_DIR$/Github-ignore/unified dataset/millionsongsubset/MillionSongSubset/A/S/E" />
+      <excludeFolder url="file://$MODULE_DIR$/Github-ignore/unified dataset/unified_data" />
+      <excludeFolder url="file://$MODULE_DIR$/venv" />
+    </content>
+    <orderEntry type="jdk" jdkName="Python 3.12 (CS467-Top-n-Music-Genre-Classification-Neural-Network)" jdkType="Python SDK" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+  <component name="PyDocumentationSettings">
+    <option name="format" value="PLAIN" />
+    <option name="myDocStringFormat" value="Plain" />
+  </component>
+</module>
